Switch imports and packages to com.samm8g.chares
