--- app/src/main/java/com/example/chares/data/AppDatabase.kt
@@ -1,38 +1,0 @@
-package com.example.chares.data
-
-import android.content.Context
-import androidx.room.Database
-import androidx.room.Room
-import androidx.room.RoomDatabase
-import androidx.room.migration.Migration
-import androidx.sqlite.db.SupportSQLiteDatabase
-
-@Database(entities = [Chore::class], version = 2, exportSchema = false)
-abstract class AppDatabase : RoomDatabase() {
-    abstract fun choreDao(): ChoreDao
-
-    companion object {
-        @Volatile
-        private var INSTANCE: AppDatabase? = null
-
-        fun getDatabase(context: Context): AppDatabase {
-            return INSTANCE ?: synchronized(this) {
-                val instance = Room.databaseBuilder(
-                    context.applicationContext,
-                    AppDatabase::class.java,
-                    "chore_database"
-                )
-                .addMigrations(MIGRATION_1_2)
-                .build()
-                INSTANCE = instance
-                instance
-            }
-        }
-
-        private val MIGRATION_1_2 = object : Migration(1, 2) {
-            override fun migrate(database: SupportSQLiteDatabase) {
-                database.execSQL("ALTER TABLE chores ADD COLUMN completedAt INTEGER")
-            }
-        }
-    }
-}
