--- app/src/main/java/com/example/chares/data/preferences/SettingsDataStore.kt
@@ -1,8 +1,0 @@
-package com.example.chares.data.preferences
-
-import android.content.Context
-import androidx.datastore.core.DataStore
-import androidx.datastore.preferences.core.Preferences
-import androidx.datastore.preferences.preferencesDataStore
-
-val Context.dataStore: DataStore<Preferences> by preferencesDataStore(name = "settings")
