--- app/src/main/java/com/example/chares/repositories/ChoreRepository.kt
@@ -1,34 +1,0 @@
-package com.example.chares.repositories
-
-import com.example.chares.data.Chore
-import com.example.chares.data.ChoreDao
-import kotlinx.coroutines.flow.Flow
-import kotlinx.serialization.json.Json
-import kotlinx.serialization.builtins.ListSerializer
-
-class ChoreRepository(private val choreDao: ChoreDao) {
-    val allChores: Flow<List<Chore>> = choreDao.getAllChores()
-
-    suspend fun insert(chore: Chore) {
-        choreDao.insertChore(chore)
-    }
-
-    suspend fun update(chore: Chore) {
-        choreDao.updateChore(chore)
-    }
-
-    suspend fun delete(chore: Chore) {
-        choreDao.deleteChore(chore)
-    }
-
-    suspend fun exportChoreData(): String {
-        val chores = choreDao.getAllChoresOnce()
-        return Json.encodeToString(ListSerializer(Chore.serializer()), chores)
-    }
-
-    suspend fun importChoreData(jsonString: String) {
-        val chores = Json.decodeFromString(ListSerializer(Chore.serializer()), jsonString)
-        choreDao.deleteAllChores()
-        choreDao.insertAll(chores)
-    }
-}
