--- app/src/main/java/com/example/chares/viewmodels/ChoreViewModel.kt
@@ -1,40 +1,0 @@
-package com.example.chares.viewmodels
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import androidx.lifecycle.viewModelScope
-import com.example.chares.data.Chore
-import com.example.chares.repositories.ChoreRepository
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.launch
-
-class ChoreViewModel(private val repository: ChoreRepository) : ViewModel() {
-
-    val allChores: Flow<List<Chore>> = repository.allChores
-
-    fun insert(chore: Chore) = viewModelScope.launch {
-        repository.insert(chore)
-    }
-
-    fun update(chore: Chore, isCompleted: Boolean) = viewModelScope.launch {
-        val updatedChore = chore.copy(
-            isCompleted = isCompleted,
-            completedAt = if (isCompleted) System.currentTimeMillis() else null
-        )
-        repository.update(updatedChore)
-    }
-
-    fun delete(chore: Chore) = viewModelScope.launch {
-        repository.delete(chore)
-    }
-}
-
-class ChoreViewModelFactory(private val repository: ChoreRepository) : ViewModelProvider.Factory {
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if (modelClass.isAssignableFrom(ChoreViewModel::class.java)) {
-            @Suppress("UNCHECKED_CAST")
-            return ChoreViewModel(repository) as T
-        }
-        throw IllegalArgumentException("Unknown ViewModel class")
-    }
-}
