+++ app/src/main/java/com/samm8g/chares/data/Chore.kt
@@ -1,0 +1,18 @@
+package com.samm8g.chares.data
+
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+import kotlinx.serialization.InternalSerializationApi
+import kotlinx.serialization.Serializable
+
+@OptIn(InternalSerializationApi::class)
+@Serializable
+@Entity(tableName = "chores")
+data class Chore(
+    @PrimaryKey(autoGenerate = true)
+    val id: Int = 0,
+    val title: String,
+    val description: String? = null,
+    val isCompleted: Boolean = false,
+    val completedAt: Long? = null
+)
