+++ app/src/main/java/com/samm8g/chares/data/preferences/SettingsDataStore.kt
@@ -1,0 +1,8 @@
+package com.samm8g.chares.data.preferences
+
+import android.content.Context
+import androidx.datastore.core.DataStore
+import androidx.datastore.preferences.core.Preferences
+import androidx.datastore.preferences.preferencesDataStore
+
+val Context.dataStore: DataStore<Preferences> by preferencesDataStore(name = "settings")
