+++ app/src/main/java/com/samm8g/chares/ui/theme/Color.kt
@@ -1,0 +1,218 @@
+package com.samm8g.chares.ui.theme
+import androidx.compose.ui.graphics.Color
+
+val primaryLight = Color(0xFF3F6836)
+val onPrimaryLight = Color(0xFFFFFFFF)
+val primaryContainerLight = Color(0xFFBFEFB0)
+val onPrimaryContainerLight = Color(0xFF275021)
+val secondaryLight = Color(0xFF53634E)
+val onSecondaryLight = Color(0xFFFFFFFF)
+val secondaryContainerLight = Color(0xFFD7E8CD)
+val onSecondaryContainerLight = Color(0xFF3C4B37)
+val tertiaryLight = Color(0xFF386569)
+val onTertiaryLight = Color(0xFFFFFFFF)
+val tertiaryContainerLight = Color(0xFFBCEBEF)
+val onTertiaryContainerLight = Color(0xFF1E4D51)
+val errorLight = Color(0xFFBA1A1A)
+val onErrorLight = Color(0xFFFFFFFF)
+val errorContainerLight = Color(0xFFFFDAD6)
+val onErrorContainerLight = Color(0xFF93000A)
+val backgroundLight = Color(0xFFF8FBF1)
+val onBackgroundLight = Color(0xFF191D17)
+val surfaceLight = Color(0xFFF8FBF1)
+val onSurfaceLight = Color(0xFF191D17)
+val surfaceVariantLight = Color(0xFFDFE4D8)
+val onSurfaceVariantLight = Color(0xFF43483F)
+val outlineLight = Color(0xFF73796E)
+val outlineVariantLight = Color(0xFFC3C8BC)
+val scrimLight = Color(0xFF000000)
+val inverseSurfaceLight = Color(0xFF2E322B)
+val inverseOnSurfaceLight = Color(0xFFEFF2E8)
+val inversePrimaryLight = Color(0xFFA4D396)
+val surfaceDimLight = Color(0xFFD8DBD2)
+val surfaceBrightLight = Color(0xFFF8FBF1)
+val surfaceContainerLowestLight = Color(0xFFFFFFFF)
+val surfaceContainerLowLight = Color(0xFFF2F5EB)
+val surfaceContainerLight = Color(0xFFECEFE5)
+val surfaceContainerHighLight = Color(0xFFE6E9E0)
+val surfaceContainerHighestLight = Color(0xFFE1E4DA)
+
+val primaryLightMediumContrast = Color(0xFF163E12)
+val onPrimaryLightMediumContrast = Color(0xFFFFFFFF)
+val primaryContainerLightMediumContrast = Color(0xFF4D7744)
+val onPrimaryContainerLightMediumContrast = Color(0xFFFFFFFF)
+val secondaryLightMediumContrast = Color(0xFF2C3A27)
+val onSecondaryLightMediumContrast = Color(0xFFFFFFFF)
+val secondaryContainerLightMediumContrast = Color(0xFF62715C)
+val onSecondaryContainerLightMediumContrast = Color(0xFFFFFFFF)
+val tertiaryLightMediumContrast = Color(0xFF073C40)
+val onTertiaryLightMediumContrast = Color(0xFFFFFFFF)
+val tertiaryContainerLightMediumContrast = Color(0xFF477478)
+val onTertiaryContainerLightMediumContrast = Color(0xFFFFFFFF)
+val errorLightMediumContrast = Color(0xFF740006)
+val onErrorLightMediumContrast = Color(0xFFFFFFFF)
+val errorContainerLightMediumContrast = Color(0xFFCF2C27)
+val onErrorContainerLightMediumContrast = Color(0xFFFFFFFF)
+val backgroundLightMediumContrast = Color(0xFFF8FBF1)
+val onBackgroundLightMediumContrast = Color(0xFF191D17)
+val surfaceLightMediumContrast = Color(0xFFF8FBF1)
+val onSurfaceLightMediumContrast = Color(0xFF0E120D)
+val surfaceVariantLightMediumContrast = Color(0xFFDFE4D8)
+val onSurfaceVariantLightMediumContrast = Color(0xFF32382F)
+val outlineLightMediumContrast = Color(0xFF4E544B)
+val outlineVariantLightMediumContrast = Color(0xFF696F64)
+val scrimLightMediumContrast = Color(0xFF000000)
+val inverseSurfaceLightMediumContrast = Color(0xFF2E322B)
+val inverseOnSurfaceLightMediumContrast = Color(0xFFEFF2E8)
+val inversePrimaryLightMediumContrast = Color(0xFFA4D396)
+val surfaceDimLightMediumContrast = Color(0xFFC4C8BF)
+val surfaceBrightLightMediumContrast = Color(0xFFF8FBF1)
+val surfaceContainerLowestLightMediumContrast = Color(0xFFFFFFFF)
+val surfaceContainerLowLightMediumContrast = Color(0xFFF2F5EB)
+val surfaceContainerLightMediumContrast = Color(0xFFE6E9E0)
+val surfaceContainerHighLightMediumContrast = Color(0xFFDBDED5)
+val surfaceContainerHighestLightMediumContrast = Color(0xFFD0D3C9)
+
+val primaryLightHighContrast = Color(0xFF0B3408)
+val onPrimaryLightHighContrast = Color(0xFFFFFFFF)
+val primaryContainerLightHighContrast = Color(0xFF2A5223)
+val onPrimaryContainerLightHighContrast = Color(0xFFFFFFFF)
+val secondaryLightHighContrast = Color(0xFF22301E)
+val onSecondaryLightHighContrast = Color(0xFFFFFFFF)
+val secondaryContainerLightHighContrast = Color(0xFF3F4D39)
+val onSecondaryContainerLightHighContrast = Color(0xFFFFFFFF)
+val tertiaryLightHighContrast = Color(0xFF003235)
+val onTertiaryLightHighContrast = Color(0xFFFFFFFF)
+val tertiaryContainerLightHighContrast = Color(0xFF215053)
+val onTertiaryContainerLightHighContrast = Color(0xFFFFFFFF)
+val errorLightHighContrast = Color(0xFF600004)
+val onErrorLightHighContrast = Color(0xFFFFFFFF)
+val errorContainerLightHighContrast = Color(0xFF98000A)
+val onErrorContainerLightHighContrast = Color(0xFFFFFFFF)
+val backgroundLightHighContrast = Color(0xFFF8FBF1)
+val onBackgroundLightHighContrast = Color(0xFF191D17)
+val surfaceLightHighContrast = Color(0xFFF8FBF1)
+val onSurfaceLightHighContrast = Color(0xFF000000)
+val surfaceVariantLightHighContrast = Color(0xFFDFE4D8)
+val onSurfaceVariantLightHighContrast = Color(0xFF000000)
+val outlineLightHighContrast = Color(0xFF282E25)
+val outlineVariantLightHighContrast = Color(0xFF454B41)
+val scrimLightHighContrast = Color(0xFF000000)
+val inverseSurfaceLightHighContrast = Color(0xFF2E322B)
+val inverseOnSurfaceLightHighContrast = Color(0xFFFFFFFF)
+val inversePrimaryLightHighContrast = Color(0xFFA4D396)
+val surfaceDimLightHighContrast = Color(0xFFB7BAB1)
+val surfaceBrightLightHighContrast = Color(0xFFF8FBF1)
+val surfaceContainerLowestLightHighContrast = Color(0xFFFFFFFF)
+val surfaceContainerLowLightHighContrast = Color(0xFFEFF2E8)
+val surfaceContainerLightHighContrast = Color(0xFFE1E4DA)
+val surfaceContainerHighLightHighContrast = Color(0xFFD2D6CC)
+val surfaceContainerHighestLightHighContrast = Color(0xFFC4C8BF)
+
+val primaryDark = Color(0xFFA4D396)
+val onPrimaryDark = Color(0xFF10380C)
+val primaryContainerDark = Color(0xFF275021)
+val onPrimaryContainerDark = Color(0xFFBFEFB0)
+val secondaryDark = Color(0xFFBBCBB2)
+val onSecondaryDark = Color(0xFF263422)
+val secondaryContainerDark = Color(0xFF3C4B37)
+val onSecondaryContainerDark = Color(0xFFD7E8CD)
+val tertiaryDark = Color(0xFFA0CFD2)
+val onTertiaryDark = Color(0xFF00373A)
+val tertiaryContainerDark = Color(0xFF1E4D51)
+val onTertiaryContainerDark = Color(0xFFBCEBEF)
+val errorDark = Color(0xFFFFB4AB)
+val onErrorDark = Color(0xFF690005)
+val errorContainerDark = Color(0xFF93000A)
+val onErrorContainerDark = Color(0xFFFFDAD6)
+val backgroundDark = Color(0xFF11140F)
+val onBackgroundDark = Color(0xFFE1E4DA)
+val surfaceDark = Color(0xFF11140F)
+val onSurfaceDark = Color(0xFFE1E4DA)
+val surfaceVariantDark = Color(0xFF43483F)
+val onSurfaceVariantDark = Color(0xFFC3C8BC)
+val outlineDark = Color(0xFF8D9387)
+val outlineVariantDark = Color(0xFF43483F)
+val scrimDark = Color(0xFF000000)
+val inverseSurfaceDark = Color(0xFFE1E4DA)
+val inverseOnSurfaceDark = Color(0xFF2E322B)
+val inversePrimaryDark = Color(0xFF3F6836)
+val surfaceDimDark = Color(0xFF11140F)
+val surfaceBrightDark = Color(0xFF363A34)
+val surfaceContainerLowestDark = Color(0xFF0B0F0A)
+val surfaceContainerLowDark = Color(0xFF191D17)
+val surfaceContainerDark = Color(0xFF1D211B)
+val surfaceContainerHighDark = Color(0xFF272B25)
+val surfaceContainerHighestDark = Color(0xFF32362F)
+
+val primaryDarkMediumContrast = Color(0xFFB9E9AB)
+val onPrimaryDarkMediumContrast = Color(0xFF032D03)
+val primaryContainerDarkMediumContrast = Color(0xFF709C64)
+val onPrimaryContainerDarkMediumContrast = Color(0xFF000000)
+val secondaryDarkMediumContrast = Color(0xFFD0E1C7)
+val onSecondaryDarkMediumContrast = Color(0xFF1C2918)
+val secondaryContainerDarkMediumContrast = Color(0xFF85957E)
+val onSecondaryContainerDarkMediumContrast = Color(0xFF000000)
+val tertiaryDarkMediumContrast = Color(0xFFB6E5E8)
+val onTertiaryDarkMediumContrast = Color(0xFF002B2D)
+val tertiaryContainerDarkMediumContrast = Color(0xFF6B989C)
+val onTertiaryContainerDarkMediumContrast = Color(0xFF000000)
+val errorDarkMediumContrast = Color(0xFFFFD2CC)
+val onErrorDarkMediumContrast = Color(0xFF540003)
+val errorContainerDarkMediumContrast = Color(0xFFFF5449)
+val onErrorContainerDarkMediumContrast = Color(0xFF000000)
+val backgroundDarkMediumContrast = Color(0xFF11140F)
+val onBackgroundDarkMediumContrast = Color(0xFFE1E4DA)
+val surfaceDarkMediumContrast = Color(0xFF11140F)
+val onSurfaceDarkMediumContrast = Color(0xFFFFFFFF)
+val surfaceVariantDarkMediumContrast = Color(0xFF43483F)
+val onSurfaceVariantDarkMediumContrast = Color(0xFFD8DED2)
+val outlineDarkMediumContrast = Color(0xFFAEB4A8)
+val outlineVariantDarkMediumContrast = Color(0xFF8C9287)
+val scrimDarkMediumContrast = Color(0xFF000000)
+val inverseSurfaceDarkMediumContrast = Color(0xFFE1E4DA)
+val inverseOnSurfaceDarkMediumContrast = Color(0xFF272B25)
+val inversePrimaryDarkMediumContrast = Color(0xFF295122)
+val surfaceDimDarkMediumContrast = Color(0xFF11140F)
+val surfaceBrightDarkMediumContrast = Color(0xFF42463F)
+val surfaceContainerLowestDarkMediumContrast = Color(0xFF050804)
+val surfaceContainerLowDarkMediumContrast = Color(0xFF1B1F19)
+val surfaceContainerDarkMediumContrast = Color(0xFF252923)
+val surfaceContainerHighDarkMediumContrast = Color(0xFF30342D)
+val surfaceContainerHighestDarkMediumContrast = Color(0xFF3B3F38)
+
+val primaryDarkHighContrast = Color(0xFFCDFDBD)
+val onPrimaryDarkHighContrast = Color(0xFF000000)
+val primaryContainerDarkHighContrast = Color(0xFFA0CF93)
+val onPrimaryContainerDarkHighContrast = Color(0xFF000F00)
+val secondaryDarkHighContrast = Color(0xFFE4F5DA)
+val onSecondaryDarkHighContrast = Color(0xFF000000)
+val secondaryContainerDarkHighContrast = Color(0xFFB7C8AE)
+val onSecondaryContainerDarkHighContrast = Color(0xFF030E02)
+val tertiaryDarkHighContrast = Color(0xFFC9F9FC)
+val onTertiaryDarkHighContrast = Color(0xFF000000)
+val tertiaryContainerDarkHighContrast = Color(0xFF9CCBCE)
+val onTertiaryContainerDarkHighContrast = Color(0xFF000E0F)
+val errorDarkHighContrast = Color(0xFFFFECE9)
+val onErrorDarkHighContrast = Color(0xFF000000)
+val errorContainerDarkHighContrast = Color(0xFFFFAEA4)
+val onErrorContainerDarkHighContrast = Color(0xFF220001)
+val backgroundDarkHighContrast = Color(0xFF11140F)
+val onBackgroundDarkHighContrast = Color(0xFFE1E4DA)
+val surfaceDarkHighContrast = Color(0xFF11140F)
+val onSurfaceDarkHighContrast = Color(0xFFFFFFFF)
+val surfaceVariantDarkHighContrast = Color(0xFF43483F)
+val onSurfaceVariantDarkHighContrast = Color(0xFFFFFFFF)
+val outlineDarkHighContrast = Color(0xFFECF2E5)
+val outlineVariantDarkHighContrast = Color(0xFFBFC5B8)
+val scrimDarkHighContrast = Color(0xFF000000)
+val inverseSurfaceDarkHighContrast = Color(0xFFE1E4DA)
+val inverseOnSurfaceDarkHighContrast = Color(0xFF000000)
+val inversePrimaryDarkHighContrast = Color(0xFF295122)
+val surfaceDimDarkHighContrast = Color(0xFF11140F)
+val surfaceBrightDarkHighContrast = Color(0xFF4D514A)
+val surfaceContainerLowestDarkHighContrast = Color(0xFF000000)
+val surfaceContainerLowDarkHighContrast = Color(0xFF1D211B)
+val surfaceContainerDarkHighContrast = Color(0xFF2E322B)
+val surfaceContainerHighDarkHighContrast = Color(0xFF393D36)
+val surfaceContainerHighestDarkHighContrast = Color(0xFF444841)
